Furnaces, Organization, Multiple outputs
Implemented Furnaces and its related classes, for now the only block using it is the animator. Organized and fixed bugs in the ArcaneTable, Recipes (both crafting and furnace) and menus/screens. Recipes now support multiple types of output, as long as the block entity uses the new functions. Inventories of block entities now save.